moved server logic to server folder
